// smsWithMavenForntEnd/src/app/model/result.model.ts
export class ResultAddModel {
    rid?: number;
    rsession?: string;
    rclass?: string;
    rbatchid?: string;
    rroll?: string;
    rbangla?: number;
    rmath?: number;
    renglish?: number;
    rislam?: number;
    rscince?: number;
    rsocial?: number;
    rtotalmark?: number;
    ravg?: number;
    rgpa?: number;
    rgrade?: string;
    rpassFail?: string;
    rexamcatagory?: string;
    // studentAddModel: StudentAddModel;
}